+++ src/eval.rs
@@ -5,56 +5,56 @@ use failure::{bail, format_err};
 use std::collections::HashMap;
 
 struct Eval<'a> {
-    env: &'a HashMap<String, isize>,
+    env: &'a HashMap<String, i64>,
 }
 
 impl<'a> AbstractInterpret for Eval<'a> {
-    type Output = isize;
+    type Output = i64;
 
-    fn constant(&mut self, c: isize) -> isize {
+    fn constant(&mut self, c: i64) -> i64 {
         c
     }
 
-    fn add(&mut self, lhs: &isize, rhs: &isize) -> isize {
+    fn add(&mut self, lhs: &i64, rhs: &i64) -> i64 {
         lhs + rhs
     }
 
-    fn sub(&mut self, lhs: &isize, rhs: &isize) -> isize {
+    fn sub(&mut self, lhs: &i64, rhs: &i64) -> i64 {
         lhs - rhs
     }
 
-    fn mul(&mut self, lhs: &isize, rhs: &isize) -> isize {
+    fn mul(&mut self, lhs: &i64, rhs: &i64) -> i64 {
         lhs * rhs
     }
 
-    fn div(&mut self, lhs: &isize, rhs: &isize) -> Result<isize> {
+    fn div(&mut self, lhs: &i64, rhs: &i64) -> Result<i64> {
         if *rhs == 0 {
             bail!("divide by zero");
         }
         Ok(lhs / rhs)
     }
 
-    fn shr(&mut self, lhs: &isize, rhs: &isize) -> isize {
+    fn shr(&mut self, lhs: &i64, rhs: &i64) -> i64 {
         lhs >> rhs
     }
 
-    fn shl(&mut self, lhs: &isize, rhs: &isize) -> isize {
+    fn shl(&mut self, lhs: &i64, rhs: &i64) -> i64 {
         lhs << rhs
     }
 
-    fn neg(&mut self, e: &isize) -> isize {
+    fn neg(&mut self, e: &i64) -> i64 {
         -e
     }
 
-    fn eq(&mut self, lhs: &isize, rhs: &isize) -> isize {
-        (lhs == rhs) as isize
+    fn eq(&mut self, lhs: &i64, rhs: &i64) -> i64 {
+        (lhs == rhs) as i64
     }
 
-    fn neq(&mut self, lhs: &isize, rhs: &isize) -> isize {
-        (lhs != rhs) as isize
+    fn neq(&mut self, lhs: &i64, rhs: &i64) -> i64 {
+        (lhs != rhs) as i64
     }
 
-    fn lookup(&mut self, var: &str) -> Result<isize> {
+    fn lookup(&mut self, var: &str) -> Result<i64> {
         self.env
             .get(var)
             .cloned()
@@ -62,7 +62,7 @@ impl<'a> AbstractInterpret for Eval<'a> {
     }
 }
 
-pub fn eval(ctx: &mut ast::Context, node: NodeId, env: &HashMap<String, isize>) -> Result<isize> {
+pub fn eval(ctx: &mut ast::Context, node: NodeId, env: &HashMap<String, i64>) -> Result<i64> {
     let eval = &mut Eval { env };
     interpret(eval, ctx, node)
 }
